Many modification
Modifications in order:
- Change of the name of the software to mieze_python as it was made clear that this software will tackle the needs of MIEZE only
- Refactoring of the fit file structure likely to be continued to allow more readability. 
- minor bug fixes that were spotted by trying to reduce a new dataset (more to come to remove hardcoded variables).
