+++ Examples/result.txt
@@ -1,15 +1,15 @@
-2018-10-05 11:30:17.719355 INFORMATION Added the entry'Date' to the metadata
-2018-10-05 11:30:17.719357 INFORMATION Added the entry'Start' to the metadata
-2018-10-05 11:30:17.719358 INFORMATION Added the entry'name' to the metadata
-2018-10-05 11:30:17.728505 INFORMATION Added the entry'Gamma' to the results
-2018-10-05 11:30:17.728511 INFORMATION Added the entry'Gamma_error' to the results
-2018-10-05 11:30:17.728513 INFORMATION Added the entry'Curve' to the results
-2018-10-05 11:30:17.728514 INFORMATION Added the entry'Parameters' to the results
-2018-10-05 11:30:17.728516 INFORMATION Added the entry'Select' to the results
-2018-10-05 11:30:17.728518 INFORMATION Added the entry'BG' to the results
-2018-10-05 11:30:17.728520 INFORMATION Added the entry'Reference' to the results
-2018-10-05 11:30:17.728521 INFORMATION Added the entry'Axis' to the results
-2018-10-05 11:30:17.728523 INFORMATION Added the entry'Axis_unit' to the results
-2018-10-05 11:30:17.728525 INFORMATION Fitting of the contrast was a success
-2018-10-05 11:30:17.728528 INFORMATION Added the entry'End' to the metadata
-2018-10-05 11:30:17.728534 INFORMATION Added the entry'Duration' to the metadata
+2018-10-22 10:01:00.826555 INFORMATION Added the entry'Date' to the metadata
+2018-10-22 10:01:00.826558 INFORMATION Added the entry'Start' to the metadata
+2018-10-22 10:01:00.826559 INFORMATION Added the entry'name' to the metadata
+2018-10-22 10:01:00.835615 INFORMATION Added the entry'Gamma' to the results
+2018-10-22 10:01:00.835620 INFORMATION Added the entry'Gamma_error' to the results
+2018-10-22 10:01:00.835622 INFORMATION Added the entry'Curve' to the results
+2018-10-22 10:01:00.835624 INFORMATION Added the entry'Parameters' to the results
+2018-10-22 10:01:00.835625 INFORMATION Added the entry'Select' to the results
+2018-10-22 10:01:00.835627 INFORMATION Added the entry'BG' to the results
+2018-10-22 10:01:00.835628 INFORMATION Added the entry'Reference' to the results
+2018-10-22 10:01:00.835630 INFORMATION Added the entry'Axis' to the results
+2018-10-22 10:01:00.835631 INFORMATION Added the entry'Axis_unit' to the results
+2018-10-22 10:01:00.835633 INFORMATION Fitting of the contrast was a success
+2018-10-22 10:01:00.835636 INFORMATION Added the entry'End' to the metadata
+2018-10-22 10:01:00.835641 INFORMATION Added the entry'Duration' to the metadata
