+++ mieze_python/core/fit_modules/fit_mieze.py
@@ -29,8 +29,7 @@ import scipy
 import warnings
 import copy
 import math
-import timeit
- 
+
 from scipy import special as sp
 from scipy import integrate as integrate
 from scipy import special as sp
@@ -38,306 +37,12 @@ from scipy import constants as co
 from scipy import optimize as op
 from scipy import stats as st
 
-#############################
-#import child components
-from .log import Log_Handler
-    
-def get_fit_handler(select):
-    '''
-    ##############################################
-    Will return the right fit manager depending 
-    on the initial input
-    ———————
-    Input: target (Data_Structure)
-    ———————
-    Output: -
-    ———————
-    status: active
-    ##############################################
-    '''
 
-    if select == 'MIEZE':
+from .fit_general import Fit_Handler
 
-        return Fit_MIEZE()
-
-    if select == 'SANS':
-
-        return Fit_SANS()
-
-    else:
-
-        print('Could not find the fit class you are looking for. Error...')
-        
-        return None
-
-
-class Fit_Handler():
-
-    def __init__(self):
-        '''
-        ##############################################
-        This is the initializer of the fit class
-        within.
-        ———————
-        Input: target (Data_Structure)
-        ———————
-        Output: -
-        ———————
-        status: active
-        ##############################################
-        '''
-        self.fun_dict   = {}
-        self.ptr_dict   = {}
-        self.para_dict  = {}
-        self.log        = Log_Handler()
-        self.verbose    = False
-
-    def __getitem__(self, key):
-        '''
-        ##############################################
-        This getitem method will be transfered to the
-        children and is here to manage different calls
-
-        key = 'error', 'warning', 'info' will return
-        the element of the log. 
-
-        key = 'result' will return the last result
-
-        key = 'print_result' will print it
-
-        key = 'results' or 'logs' will return the 
-        actual classes
-
-        any other key will try to grab the dictionary
-        ———————
-        Input: target (Data_Structure)
-        ———————
-        Output: -
-        ———————
-        status: active
-        ##############################################
-        '''
-        if key == 'error':
-
-            return self.log.return_last_log('error')
-
-        elif key == 'info':
-
-            return self.log.return_last_log('info')
-
-        elif key == 'warning':
-
-            return self.log.return_last_log('warning')
-
-        else:
-
-            return self.fun_dict[key]
-
-
-    def set_method(self, target, identifier):
-        '''
-        ##############################################
-        This fucntion will try to select the right
-        pointer for the right method
-        ———————
-        Input: 
-        - target (str) the target method keyword
-        - identifier (str) the method to select
-        ———————
-        Output: -
-        ———————
-        status: active
-        ##############################################
-        '''
-        if target in self.fun_dict and identifier in self.ptr_dict:
-
-            self.fun_dict[target]= self.ptr_dict[identifier]
-
-        else:
-
-            print('The input keys do not match existing functions')
-
-    def set_parameter(self, name = '', value = ''):
-        '''
-        ##############################################
-        This function will allow the user to inject
-        fit parameters...
-        ———————
-        Input: -
-        ———————
-        Output: -
-        ———————
-        status: active
-        ##############################################
-        '''
-        self.para_dict[name] = value
-
-class Fit_SANS(Fit_Handler):
-
-    def __init__(self):
-        '''
-        ##############################################
-        This is the initializer of the SANS fit class
-        within. It will also initialize the superclass
-        containing the generalized methods.
-        ———————
-        Input: -
-        ———————
-        Output: -
-        ———————
-        status: active
-        ##############################################
-        '''
-        #initialize the superclass
-        Fit_Handler.__init__(self)
-
-        self.ptr_dict = {}
-        self.ptr_dict['intensity']      = self.intensity_vs_parameter
-        self.set_defaults()
-        self.set_fit_parameters()
-
-    def set_defaults(self):
-        '''
-        ##############################################
-        This function will build the default 
-        dictionary of function dict that will linkt 
-        the functions to the selected method
-        ———————
-        Input: -
-        ———————
-        Output: -
-        ———————
-        status: active
-        ##############################################
-        '''
-
-        self.fun_dict = {}
-        self.fun_dict['intensity']     = self.ptr_dict['intensity']
-
-    def set_fit_parameters(self):
-        '''
-        ##############################################
-        This function will build the default 
-        dictionary of function dict that will linkt 
-        the functions to the selected method
-        ———————
-        Input: -
-        ———————
-        Output: -
-        ———————
-        status: active
-        ##############################################
-        '''
-        ############################################
-        #pack them into the dictionary
-        self.para_dict = {}
-        self.para_dict['BG substraction']   = None
-
-    def intensity_vs_parameter(self, target, mask, results):
-        '''
-        ##############################################
-        This method aims atprocessing the SANS
-        intensity against the a given parameter
-        ———————
-        Input: -
-        ———————
-        Output: -
-        ———————
-        status: active
-        ##############################################
-        '''
-        ##############################################
-        #Initialize the output dictionary with all def.
-        local_results = results.generate_result( name = 'Intensity')
-
-        ##############################################
-        #Initialize the output dictionary with all def.
-        mask            = mask.mask
-        intensity       = {}
-        intensity_error = {}
-        axis            = {}
-
-        ##############################################
-        #process BG metadata if necessary
-        if not self.para_dict['BG substraction'] == None:
-
-            BG_target = target.get_slice([self.para_dict['BG substraction']])
-
-            monitor_BG = BG_target.get_metadata([0])[0]['monitor']
-
-        ##############################################
-        #loop over and process
-        for idx_0 in range(len(target.axes.axes[0])):
-            
-            #set the key for data saving
-            key  = target.axes.axes[0][idx_0]
-
-            #process the data slice
-            new_target = target.get_slice([key])
-            data       = new_target.return_as_np()
-
-            #set the axis
-            axis[key]               = new_target.axes.axes[0]
-            intensity[key]          = np.zeros(len(new_target.axes.axes[0]))
-            intensity_error[key]    = np.zeros(len(new_target.axes.axes[0]))
-
-            for idx_1 in range(new_target.axes.axes_len[0]):
-
-                #grab the metadata
-                monitor = new_target.get_metadata([idx_1])[0]['monitor']
-
-                if self.para_dict['BG substraction'] == None:
-
-                    #process the instensity
-                    intensity[key][idx_1] = np.sum(data[idx_1]*mask)/monitor/np.sum(mask)
-
-                    #process its error
-                    intensity_error[key][idx_1] = np.sqrt(np.sum(data[idx_1]*mask))/monitor/np.sum(mask)
-
-                else:
-                
-                    if data[0].shape[0] == data[idx_1].shape[0]:
-                        
-                        #process the instensity
-                        intensity[key][idx_1] = np.sum(data[idx_1]*mask)/monitor/np.sum(mask) - np.sum(BG_target[0]*mask)/monitor_BG/np.sum(mask)
-
-                        #process its error
-                        intensity_error[key][idx_1] = np.sqrt(np.sum(data[idx_1]*mask)/monitor**2/np.sum(mask)**2 + np.sum(BG_target[0]*mask)/monitor_BG**2/np.sum(mask)**2)
-
-                    else:
-
-                        #process the instensity
-                        intensity[key][idx_1] = (
-                            np.sum(data[idx_1] * mask)
-                            / monitor
-                            /np.sum(mask) 
-                            - np.sum(BG_target[0]*mask)
-                            /monitor_BG
-                            /np.sum(mask))
-                        
-                        #process its error
-                        intensity_error[key][idx_1] = (np.sqrt(
-                            np.sum(data[idx_1]*mask)/monitor**2/np.sum(mask)**2 
-                            + np.sum(BG_target[0]*mask)/monitor_BG**2/np.sum(mask)**2))
-
-        
-        ##############################################
-        #finalize result and send it out
-        local_results['Intensity']       = intensity
-        local_results['Intensity_error'] = intensity_error
-        local_results['Axis']            = axis
-    
-        #write the dictionary entries
-        local_results.add_log('info', 'Computation of the intensity was a success')
-        local_results.set_complete()
-
-        #tell fit handler what happened
-        self.log.add_log(
-            'Info', 
-            'Computation of the intensity was a success')
 
 class Fit_MIEZE(Fit_Handler):
-
+    
     def __init__(self):
         '''
         ##############################################
@@ -571,8 +276,9 @@ class Fit_MIEZE(Fit_Handler):
             reference = self.para_dict['Reference']
 
             #check the dimension of this list
-            if not reference == None and not reference in target.axes.axes[0]:
-
+            if not reference == None and not reference[0] in target.axes.axes[0]:
+            
+                print(target.axes.axes[0])
                 print('This reference does not exist. Error...')
                 return False
 
@@ -669,6 +375,8 @@ class Fit_MIEZE(Fit_Handler):
 
                         #logical check
                         if foils_in_echo[echo_idx_0][foil_idx] == True:
+
+                            print(key, meas, echo, foil_idx)
                             
                             data = np.array(
                                 [(np.multiply(shift[key][meas][echo][foil_idx,tcha_idx],premask)).sum() 
@@ -691,7 +399,7 @@ class Fit_MIEZE(Fit_Handler):
 
                     ############################################
                     #!!!BUG!!!!
-                    monitor = int(1.e-3 * int((new_target.get_metadata([echo_idx_1,0,0])[0]['monitor'])))
+                    monitor = new_target.get_metadata([echo_idx_1,0,0])[0]['monitor']
 
                     ############################################
                     #process the result
@@ -832,6 +540,8 @@ class Fit_MIEZE(Fit_Handler):
             'Processing the reference contrast calculation for: '
             +str(reference)
             +target.axes.units[0])
+
+        reference = reference[0]
 
         ref_result = self.calc_contrast_fit(
             [reference], 
@@ -1519,7 +1229,7 @@ class Fit_MIEZE(Fit_Handler):
 
         ############################################
         #extract the relevant parameters
-        selected_ref    = target.metadata_class['Resolution']
+        selected_ref    = self.para_dict['Reference']
         reso_target     = target.get_slice(selected_ref)
         premask         = mask.mask
