+++ mieze_python/core/io_modules/import_mieze_tof.py
@@ -282,7 +282,10 @@ class Import_MIEZE_TOF:
 
                 target.axes.set_name(idx,name)
                 target.axes.set_unit(idx,units[idx])
-                target.axes.set_axis(idx,[float(element) for element in real_axis_definition[idx]])
+                try:
+                    target.axes.set_axis(idx,[float(element) for element in real_axis_definition[idx]])
+                except:
+                    target.axes.set_axis(idx,[element for element in real_axis_definition[idx]])
 
             except:
                 pass
