+++ mieze_python/core/result.py
