+++ trial_vis_2.txt
@@ -1,0 +1,46 @@
+
+#############################################################################
+#Specify here the full path to the data
+Path : /Users/alexanderschober/Desktop/Archive/
+
+#############################################################################
+#Specify here the data to be disregarded
+# second argument is the dimension and third either a list() or range() command
+Ignore : 3 : (3,4)
+
+#############################################################################
+#Specify here the measurement parameters
+#
+
+Dim : Temperature : Measurement : Echo : Foil : Time Channel : Pixel : Pixel
+Unit : K : - : s : - : - : px : px
+
+Data : [68.00,0] : 0012 : range(2186, 2194) : (8x16x128x128)
+Data : [31.00,0] : 0012 : range(2387, 2395) : (8x16x128x128)
+Data : [30.00,0] : 0012 : [2396, 2398, 2401, 2402, 2403] : (8x16x128x128)
+Data : [30.00,1] : 0012 : [2408, 2409, 2410] : (8x16x128x128)
+Data : [29.80,0] : 0012 : [2412, 2413, 2374, 2375, 2376, 2377, 2378, 2379] : (8x16x128x128)
+Data : [29.60,0] : 0012 : [2320, 2321, 2414, 2323, 2415, 2325, 2326, 2327] : (8x16x128x128)
+Data : [29.40,0] : 0012 : range(2334, 2342) : (8x16x128x128)
+Data : [29.20,0] : 0012 : [2348, 2349, 2350, 2351, 2352, 2353, 2354] : (8x16x128x128)
+Data : [29.20,1] : 0012 : [2416] : (8x16x128x128)
+Data : [29.10,0] : 0012 : range(2418, 2426) : (8x16x128x128)
+Data : [29.00,0] : 0012 : [2362, 2363, 2364, 2365, 2366, 2426, 2368, 2369] : (8x16x128x128)
+Data : [28.60,0] : 0012 : range(2434, 2442) : (8x16x128x128)
+Data : [28.40,0] : 0012 : range(2306, 2314) : (8x16x128x128)
+
+#############################################################################
+#Manage metadata to save
+#type : name_in_file         : name to assign in dataclass
+
+Metadata : cbox_0a_fg_freq_value : freq_0 : float : 1 : Hz
+Metadata : cbox_0b_fg_freq_value : freq_1 : float : 1 : Hz
+Metadata : selector_lambda_value : wavelength : float : 1 : A
+Metadata : psd_distance_value : lsd : float : 1 : m
+Metadata : monitor1_value : monitor : float : 1 : -
+Metadata : monitor1 : monitor : float : 1 : -
+Metadata : T_ccr_a_value : T_A : float : 1 : K
+Metadata : T_ccr_b_value : T_B : float : 1 : K
+Metadata : T_ccr_c_value : T_C : float : 1 : K
+Metadata : T_ccr_d_value : T_D : float : 1 : K
+Metadata : B_value : Bs : float : 1 : K
